move decide core to optimizely
